Move marker to dedicated file
