+++ src/routes/reviews/[category]/[slug].json.ts
@@ -1,8 +1,7 @@
 import type { Request, Response } from 'express';
-import { parseFile, aquaMark } from '../../../utils/parser';
+import { parseFile } from '../../../utils/parser';
 import { countAverageRating, contentParser } from '../../../utils/article';
-
-const mark = (data: any, str: string) => aquaMark(contentParser(data, str));
+import marker from '../../../utils/marker';
 
 export function get(req: Request, res: Response) {
 	const { category, slug } = req.params;
@@ -14,10 +13,10 @@ export function get(req: Request, res: Response) {
 		review.composed = (dStart - +new Date(data.last_seen)) / 24 / 60 / 60 / 1000;
 
 		const [article, closing] = content.split(/^## \$CLOSING/m);
-		if (closing) review.closing = mark(review, closing);
+		if (closing) review.closing = marker.render(contentParser(review, closing));
 
 		const [summary, spoilers] = article.split(/^## \$SPOILERS/m);
-		if (spoilers) review.spoilers = mark(review, spoilers);
+		if (spoilers) review.spoilers = marker.render(contentParser(review, spoilers));
 
 		review.content = contentParser(review, summary);
 		review.rating = countAverageRating(data.rating);
