+++ src/utils/marker.d.ts
@@ -1,0 +1,1 @@
+export function render(str: string): string;
